+++ tests/volume.rs
@@ -72,6 +72,36 @@ fn open_all_volumes() {
     };
 }
 
+#[test]
+fn close_volume_too_early() {
+    let time_source = utils::make_time_source();
+    let disk = utils::make_block_device(utils::DISK_SOURCE).unwrap();
+    let mut volume_mgr = embedded_sdmmc::VolumeManager::new(disk, time_source);
+
+    let volume = volume_mgr
+        .open_volume(embedded_sdmmc::VolumeIdx(0))
+        .expect("open volume 0");
+    let root_dir = volume_mgr.open_root_dir(volume).expect("open root dir");
+
+    // Dir open
+    let r = volume_mgr.close_volume(volume);
+    let Err(embedded_sdmmc::Error::VolumeStillInUse) = r else {
+        panic!("Expected failure to close volume: {r:?}");
+    };
+
+    let _test_file = volume_mgr
+        .open_file_in_dir(root_dir, "64MB.DAT", embedded_sdmmc::Mode::ReadOnly)
+        .expect("open test file");
+
+    volume_mgr.close_dir(root_dir).unwrap();
+
+    // File open, not dir open
+    let r = volume_mgr.close_volume(volume);
+    let Err(embedded_sdmmc::Error::VolumeStillInUse) = r else {
+        panic!("Expected failure to close volume: {r:?}");
+    };
+}
+
 // ****************************************************************************
 //
 // End Of File